+++ carp_mobile_sensing/lib/runtime/runtime.dart
@@ -12,6 +12,7 @@ library runtime;
 
 import 'dart:async';
 import 'dart:convert';
+import 'dart:math';
 
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:carp_core/carp_core.dart';